+++ src/main/java/at/porscheinformatik/seleniumcomponents/clarity/AbstractClarityComboboxComponent.java
@@ -3,15 +3,19 @@ package at.porscheinformatik.seleniumcomponents.clarity;
 import static at.porscheinformatik.seleniumcomponents.SeleniumAsserts.*;
 import static org.hamcrest.Matchers.*;
 
+import java.util.List;
+
 import org.openqa.selenium.Keys;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.ActiveSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponentFactory;
+import at.porscheinformatik.seleniumcomponents.SeleniumMatchers;
 import at.porscheinformatik.seleniumcomponents.SeleniumUtils;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 import at.porscheinformatik.seleniumcomponents.clarity.ClarityComboboxOptionsComponent.AbstractClarityComboboxOptionComponent;
+import at.porscheinformatik.seleniumcomponents.component.ButtonComponent;
 import at.porscheinformatik.seleniumcomponents.component.InputComponent;
 
 /**
@@ -21,6 +25,9 @@ public abstract class AbstractClarityComboboxComponent<OPTION_TYPE extends Abstr
     extends AbstractSeleniumComponent implements ActiveSeleniumComponent
 {
     protected final InputComponent input = new InputComponent(this);
+
+    protected final ButtonComponent trigger =
+        new ButtonComponent(this, WebElementSelector.selectByClassName("clr-combobox-trigger"));
 
     protected final ClarityComboboxOptionsComponent<OPTION_TYPE> options;
 
@@ -46,6 +53,20 @@ public abstract class AbstractClarityComboboxComponent<OPTION_TYPE extends Abstr
 
         // blur in order to make sure, that the options are closed
         input.type(Keys.TAB);
+    }
+
+    public List<String> getOptionLabels()
+    {
+        trigger.click();
+
+        assertThatSoon(() -> options, SeleniumMatchers.isVisible());
+
+        List<String> labels = options.getOptionLabels();
+
+        // blur in order to make sure, that the options are closed
+        input.type(Keys.TAB);
+
+        return labels;
     }
 
     public abstract void assertSelected(String partialText);